01_etl/requirements.txt
flake8==4.0.1
psycopg2-binary==2.9.3
redis==4.2.2
python-dateutil==2.8.2
pydantic==1.9.0
